+++ QA.md
@@ -1,48 +1,45 @@
-QA Testing 
-======
+# QA
 
-## Looking Glass Components: 
-
-###Sign Up Form: 
-  1. Inputs ( checking for validation + proper submit functionality): 
-    * First Name 
-    * Last Name 
-    * Email (Validation)  
-    * Password 
-    * Company 
-    * Title 
-  2. Does Sign Up Button trigger submit event?
-  3. Does Submit properly create an account with Ajax call?
-  4. Do errors come back from the AJAX call?
-  5. Google Analytics Sign Up Event Working
+## Sign Up Form: 
+1. Inputs ( checking for validation + proper submit functionality): 
+  * First Name 
+  * Last Name 
+  * Email (Validation)  
+  * Password 
+  * Company 
+  * Title 
+1. Does Sign Up Button trigger submit event?
+1. Does Submit properly create an account with Ajax call?
+1. Do errors come back from the AJAX call?
+1. Google Analytics Sign Up Event Working
  
-### Notification:
-  1. Does component appear at the proper time? 
-  2. Is the message accurate to the current error?
+## Message:
+1. Does component appear at the proper time? 
+1. Is the message accurate to the current error?
 
-### Thumbnail: 
-  1. Does Thumbnail appear?
-  2. Is checkbox functional for the thumbnail?
+## Thumbnail: 
+1. Does Thumbnail appear?
+1. Is checkbox functional for the thumbnail?
 
-### Upload Form: 
-  1. Inputs: 
-    * URL ( checking for validation + proper submit functionality)?
-  2. Upload Button?
-  3. CheckBox disable Button?
-  4. Processing Button?
-  5. Does Submit properly create an account with Ajax call?
-  6. Do errors come back from the AJAX call?
-  7. Google Analytics Upload Event Working?
+## Upload Form: 
+1. Inputs: 
+  * URL ( checking for validation + proper submit functionality)?
+1. Upload Button?
+1. CheckBox disable Button?
+1. Processing Button?
+1. Does Submit properly create an account with Ajax call?
+1. Do errors come back from the AJAX call?
+1. Google Analytics Upload Event Working?
 
-### Video:
-  1. Does serving Bar show up / Change during Progress?
-  2. Does Error Show when Progress Fails?
-  3. Do thumbnails appear for video ( max 5 )? 
-  4. Does Info Table show up?
+## Video:
+1. Does serving Bar show up / Change during Progress?
+1. Does Error Show when Progress Fails?
+1. Do thumbnails appear for video ( max 5 )? 
+1. Does Info Table show up?
 
-### Video Info Table: 
-  1. Does Correct ID appear?
-  2. Is Duration Calculated correctly + format?
-  3. Is published calculated correctly + format?
-  4. Does Source correctly link back to video ?
+## Video Info Table: 
+1. Does Correct ID appear?
+1. Is Duration Calculated correctly + format?
+1. Is published calculated correctly + format?
+1. Does Source correctly link back to video ?
