+++ src/js/components/core/App.js
@@ -5,7 +5,7 @@ import React from 'react';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - 
 
-import CallToAction from '../lookingglass/CallToAction';
+import CallToAction from '../wonderland/CallToAction';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
