+++ src/js/components/wonderland/NewsFlash.js
