+++ src/js/components/wonderland/Thumbnail.js
