+++ src/js/components/wonderland/Thumbnails.js
